backend/requirements.txt
asgiref==3.8.1
attrs==25.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
defusedxml==0.7.1
dj-rest-auth==7.0.1
Django==5.2
django-allauth==0.61.1
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
oauthlib==3.2.2
packaging==25.0
pluggy==1.5.0
pycparser==2.22
PyJWT==2.9.0
pytest==8.3.5
pytest-django==4.11.1
python-dotenv==1.1.0
python3-openid==3.2.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.24.0
sqlparse==0.5.3
uritemplate==4.1.1
urllib3==2.4.0
